add list of waypoints with indexes
