+++ rover_gui/pages/add_location_popup.py
@@ -22,7 +22,6 @@ class AddLocationPopup(QWidget):
         self.le_name : QLineEdit
         self.le_latitude : QLineEdit
         self.le_longitude : QLineEdit
-        self.cb_color : QComboBox
 
         self.pb_add_location.clicked.connect(self.add_location)
         self.pb_cancel.clicked.connect(self.cancel)
@@ -42,7 +41,7 @@ class AddLocationPopup(QWidget):
                 lines = []
         with open(self.nav_widget.saved_locations_path, "a") as f:
             index = len(lines) + 1 
-            f.write(f"{index};{self.le_name.text()};{self.le_latitude.text()};{self.le_longitude.text()};{self.cb_color.currentText()}\n")
+            f.write(f"{self.le_name.text()};{self.le_latitude.text()};{self.le_longitude.text()}\n")
 
         self.nav_widget.location_list.clear()
         self.nav_widget.load_locations()
